Waiter_agent/requirements.txt
a2a-sdk==0.3.19
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.5
azure-common==1.1.28
azure-core==1.34.0
azure-mgmt-core==1.5.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.5.0
charset-normalizer==3.4.2
chromadb==1.1.1
click==8.1.8
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
crewai==1.6.1
cryptography==45.0.6
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.3.1
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
durationpy==0.10
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastuuid==0.14.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.7.0
google-adk==1.19.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.40.3
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.128.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.34.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.59.0
google-cloud-speech==2.34.0
google-cloud-storage==3.6.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.52.0
google-generativeai==0.8.5
google-resumable-media==2.8.0
googleapis-common-protos==1.70.0
grandalf==0.8
graphviz==0.21
greenlet==3.2.2
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
httpx_auth==0.23.1
huggingface-hub==0.34.3
humanfriendly==10.0
identify==2.6.15
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.13.0
ipykernel==7.0.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
json5==0.12.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kubernetes==33.1.0
langchain-core==1.1.0
langchain-google-genai==3.2.0
langgraph-sdk==0.2.0
langsmith==0.4.13
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.22.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.4.4
nest-asyncio==1.6.0
networkx==3.5
nodeenv==1.9.1
numpy==2.2.5
oauthlib==3.2.2
onnxruntime==1.22.0
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.1
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.1
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pyvis==0.3.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.3
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlalchemy-spanner==1.17.1
sqlparse==0.5.4
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.10.0
tokenizers==0.21.4
tomli==2.2.1
tomli_w==1.2.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
ty==0.0.1a29
typer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uv==0.8.6
uvicorn==0.34.2
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wrapt==2.0.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0
